ai_assistant/pyproject.toml
[project]
name = "ai_assistant"
version = "0.1.0"
description = "A simple AI chat assistant using OpenAI's GPT-3.5 API"
dependencies = [
    "openai==1.3.0",
    "python-dotenv==1.0.0",
    "requests==2.31.0"
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
